+++ src/main/java/dev/hegel/Hegel.java
@@ -224,7 +224,6 @@ public class Hegel {
 
       if (isFinal) {
         System.err.println();
-        tc.printDrawnValues(System.err);
         System.err.println(e.getMessage());
         e.printStackTrace(System.err);
       }